requirements.txt
pytest>=8.0.0
requests>=2.31.0
allure-pytest>=2.15.0
pydantic>=2.6.2